+++ quoteAnime/Theme/Typography.swift
@@ -2,10 +2,10 @@ import SwiftUI
 
 extension Font {
     static func quoteSerif(size: CGFloat) -> Font {
-        .custom("Georgia", size: size)
+        .custom("Didot", size: size)
     }
 
     static func quoteSerifItalic(size: CGFloat) -> Font {
-        .custom("Georgia-Italic", size: size)
+        .custom("Georgia", size: size)
     }
 }